Refactor player selection and bot delay UI controls
- Consolidate bot delay controls to single row with text fields for base and delta values
- Change bot delay base from integer to float for more precise control
- Replace vertical player count buttons with horizontal row of number buttons (2-8)
- Replace bottom back button with back arrow icon in upper left corner
- Remove stepper component in favor of direct text input

+++ BattleZone/composeApp/src/commonMain/kotlin/com/franklinharper/battlezone/TurnCoordinator.kt
@@ -55,7 +55,7 @@ class TurnCoordinator(
         isCurrentPlayerBot: Boolean,
         gamePhase: GamePhase,
         hasBotDecision: Boolean,
-        botDelayBaseSeconds: Int,
+        botDelayBaseSeconds: Float,
         botDelayDeltaSeconds: Float
     ) {
         if (gamePhase == GamePhase.REINFORCEMENT) {
@@ -102,7 +102,7 @@ class TurnCoordinator(
         }
     }
 
-    private fun randomBotDelayMs(baseSeconds: Int, deltaSeconds: Float): Long {
+    private fun randomBotDelayMs(baseSeconds: Float, deltaSeconds: Float): Long {
         val clampedBase = baseSeconds.coerceIn(
             UiConstants.BOT_DELAY_BASE_MIN_SECONDS,
             UiConstants.BOT_DELAY_BASE_MAX_SECONDS
